requirements.txt
absl-py==0.11.0
alabaster==0.7.12
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.1
astunparse==1.6.3
atomicwrites==1.3.0
attrs==21.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.0.1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.2.0
catboost==0.24.1
certifi==2020.12.5
cffi==1.12.3
chardet==4.0.0
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.8.4
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.2
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
flatbuffers==1.12
fsspec==0.5.2
future==0.18.2
gast==0.3.3
gevent==1.4.0
gitdb==4.0.5
GitPython==3.1.12
glob2==0.7
google-api-core==1.23.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.2.0
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
graphviz==0.14.1
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.0
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==3.3.0
importlib-resources==5.2.0
inflection==0.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==3.0.1
joblib==1.0.1
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-kernel-gateway==2.4.3
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lesscpy==0.13.0
libarchive-c==2.8
lightgbm==3.0.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
menuinst==1.4.16
missingno==0.5.0
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multimethod==1.4
multipledispatch==0.6.0
mysql-connector-python==8.0.22
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.5.1
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.19.4
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
openpyxl==3.0.0
opt-einsum==3.3.0
packaging==19.2
pandas==1.3.0
pandas-profiling==3.0.0
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
phik==0.11.2
pickleshare==0.7.5
Pillow==6.2.0
pipreqs==0.4.10
pkginfo==1.5.0.1
plotly==4.10.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.6.3
py==1.8.0
pyarrow==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydantic==1.8.2
pydeck==0.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
Quandl==3.5.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.14.0
rsa==4.6
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scikit-plot==0.3.7
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
streamlit==0.84.1
streamlit-pandas-profiling==0.1.2
streamlit-player==0.1.5
sweetviz==2.1.3
sympy==1.4
tables==3.5.2
tangled-up-in-unicode==0.1.0
tblib==1.4.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.26.2
validators==0.18.2
visions==0.7.1
watchdog==1.0.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xgboost==1.1.1
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
xmltodict==0.12.0
yarg==0.1.9
zict==1.0.0
zipp==3.4.0
